LPD-91104 Rename Helpers to Senders
